app/requirements.txt
uvicorn==0.38.0 # Веб-сервера ASGI
httptools==0.7.1 # C-библиотека для парсинга HTTP (зависимость uvicorn)
fastapi[all]==0.120.3 # Веб-фреймворк для создания API
starlette==0.49.1 # Фреймворк ASGI для создания асинхронных веб-сервисов на Python
python-dotenv==1.2.1 # Чтение переменных окружения и .env
apscheduler==3.11.0  # Приложение для шедуллера запускающего события по расписнаию
SQLAlchemy==2.0.44 # Работа с БД
psycopg2-binary==2.9.11 # Расширения для работы с PostgreSQL
asyncpg==0.30.0 # Работа с асинхронностью БД
jinja2==3.1.6 # Шаблонизатор для генирациии конфигурационного файла NGINX
pydantic==2.12.3  # Проверка входных данных в POST-запросах на инстансы
pip-system-certs==5.3 # Разрешает использовать системные SSL-сертификаты
httpx==0.28.1 # Клиентская HTTP-библиотека для Python
python-ldap==3.4.5 # Для работы с запросами к DS
bcrypt==5.0.0 # Хеширование паролей

